requirements.txt
# Core AI/ML Dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
diffusers>=0.24.0
accelerate>=0.25.0

# Wan2.2 Dependencies
einops>=0.7.0
omegaconf>=2.3.0
pytorch-lightning>=2.1.0
safetensors>=0.4.1
xformers>=0.0.23

# Video Processing
opencv-python>=4.9.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3
pillow>=10.1.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Web Server
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.12.0
pylint>=3.0.0